+++ common/src/main/java/com/coolerpromc/resourcestrees/CommonClass.java
@@ -1,30 +1,22 @@
 package com.coolerpromc.resourcestrees;
 
 import com.coolerpromc.resourcestrees.block.ModBlocks;
+import com.coolerpromc.resourcestrees.block.custom.ResourcesLeavesBlock;
+import com.coolerpromc.resourcestrees.block.custom.ResourcesSaplingBlock;
 import com.coolerpromc.resourcestrees.block.entity.ModBlockEntities;
 import com.coolerpromc.resourcestrees.datacomponent.ModDataComponents;
 import com.coolerpromc.resourcestrees.item.ModCreativeTab;
 import com.coolerpromc.resourcestrees.item.ModItems;
-import com.coolerpromc.resourcestrees.platform.Services;
 import com.coolerpromc.resourcestrees.recipe.ModRecipes;
 import com.coolerpromc.resourcestrees.screen.ModMenuTypes;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.world.level.block.Block;
 
-// This class is part of the common project meaning it is shared between all supported loaders. Code written here can only
-// import and access the vanilla codebase, libraries used by vanilla, and optionally third party libraries that provide
-// common compatible binaries. This means common code can not directly use loader specific concepts such as NeoForge events
-// however it will be compatible with all supported mod loaders.
+import java.util.ArrayList;
+import java.util.List;
+
 public class CommonClass {
-
-    // The loader specific projects are able to import and use any code from the common project. This allows you to
-    // write the majority of your code here and load it from your loader specific projects. This example has some
-    // code that gets invoked by the entry point of the loader specific projects.
     public static void init() {
-
-        Constants.LOG.info("Hello from Common init on {}! we are currently in a {} environment!", Services.PLATFORM.getPlatformName(), Services.PLATFORM.getEnvironmentName());
-
-        // Trigger class loading of all registration classes.
-        // On NeoForge, Services.REGISTRY is NeoForgeRegistryHelper which uses DeferredRegister (must be registered with IEventBus before this).
-        // On Fabric, Services.REGISTRY is FabricRegistryHelper which registers directly into BuiltInRegistries.
         ModBlocks.init();
         ModItems.init();
         ModDataComponents.init();
@@ -32,5 +24,43 @@ public class CommonClass {
         ModRecipes.init();
         ModCreativeTab.init();
         ModMenuTypes.init();
+    }
+
+    public static Block[] leavesBlock(){
+        List<Block> blocks = new ArrayList<>();
+
+        for (Block block : BuiltInRegistries.BLOCK){
+            if (block instanceof ResourcesLeavesBlock) {
+                blocks.add(block);
+            }
+        }
+
+        return blocks.toArray(new Block[0]);
+    }
+
+    public static Block[] saplingBlock(){
+        List<Block> blocks = new ArrayList<>();
+
+        for (Block block : BuiltInRegistries.BLOCK){
+            if (block instanceof ResourcesSaplingBlock) {
+                blocks.add(block);
+            }
+        }
+
+        return blocks.toArray(new Block[0]);
+    }
+
+    public static List<Block> blocks(){
+        List<Block> blocks = new ArrayList<>();
+
+        for (Block block : BuiltInRegistries.BLOCK){
+            if (block instanceof ResourcesSaplingBlock || block instanceof ResourcesLeavesBlock){
+                if (block.builtInRegistryHolder().key().identifier().getNamespace().equals("kubejs")){
+                    blocks.add(block);
+                }
+            }
+        }
+
+        return blocks;
     }
 }